refactor(rating): improve handling of steps prop
