+++ packages/core/src/Rating/index.ts
@@ -1,3 +1,3 @@
-export { default as RatingGroup, type RatingGroupProps } from './RatingGroup.vue'
 export { default as RatingItem, type RatingItemProps } from './RatingItem.vue'
+export { default as RatingItemTrigger } from './RatingItemTrigger.vue'
 export { default as RatingRoot, type RatingRootProps } from './RatingRoot.vue'
