+++ packages/core/src/Rating/story/RatingDefault.story.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { Icon } from '@iconify/vue'
 import { ref } from 'vue'
-import { RatingItem, RatingRoot } from '../'
+import { RatingItem, RatingItemTrigger, RatingRoot } from '../'
 
 const value = ref(3)
 </script>
@@ -9,7 +9,7 @@ const value = ref(3)
 <template>
   <Story
     title="Rating/Demo"
-    :layout="{ type: 'single', iframe: true }"
+    :layout="{ type: 'single' }"
   >
     <Variant title="default">
       <div class="flex flex-col items-center gap-2">
@@ -19,16 +19,19 @@ const value = ref(3)
         <RatingRoot
           v-slot="{ ratings }"
           v-model="value"
-          clearable
           class="flex gap-2"
         >
           <RatingItem
             v-for="rating in ratings"
             :key="rating"
             :rating="rating"
-            class=" data-[state=active]:text-yellow-500 text-gray-400"
           >
-            <Icon icon="radix-icons:star" />
+            <RatingItemTrigger class="data-[state=active]:text-yellow-500 text-gray-400">
+              <Icon
+                icon="radix-icons:star"
+                class="size-14"
+              />
+            </RatingItemTrigger>
           </RatingItem>
         </RatingRoot>
       </div>
